# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Stellar wallet backend — auth, encrypted credential backup, and account recovery.",
        "title": "Latch API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/on-ramp/intent/{id}": {
            "get": {
                "description": "Dev-only (403 in production). Fetches an on-ramp intent by id, including its live MoonPay transaction status if one is attached.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "on-ramp"
                ],
                "summary": "Get an on-ramp intent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "On-ramp intent ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.onRampIntentResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Dev-only (403 in production). Partial update: at least one of status or moonpayTransactionId is required. Returns the updated intent with its live MoonPay transaction status.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "on-ramp"
                ],
                "summary": "Update an on-ramp intent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "On-ramp intent ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.updateOnRampIntentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.onRampIntentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/on-ramp/pool": {
            "get": {
                "description": "Dev-only (403 in production). Returns the on-ramp pool account's XLM balance and up to 20 recent transactions, optionally filtered to a single memo.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "on-ramp"
                ],
                "summary": "Get the on-ramp pool account snapshot",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter recent transactions to this memo (e.g. an intent's memoId)",
                        "name": "memo",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.onRampPoolResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/on-ramp/session": {
            "post": {
                "description": "Dev-only (403 in production). Creates a MoonPay on-ramp intent for destinationCAddress and returns either a Platform API session token or a signed widget URL depending on MOONPAY_INTEGRATION_MODE (\"auto\" falls back to a widget URL if the Platform API returns 404).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "on-ramp"
                ],
                "summary": "Create an on-ramp session",
                "parameters": [
                    {
                        "description": "Destination contract address and optional fiat amount/code",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.createOnRampSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.onRampSessionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handler_webapp.createOnRampSessionRequest": {
            "type": "object",
            "required": [
                "destinationCAddress"
            ],
            "properties": {
                "destinationCAddress": {
                    "type": "string",
                    "example": "CABC...XYZ"
                },
                "fiatAmount": {
                    "type": "string",
                    "example": "25"
                },
                "fiatCode": {
                    "type": "string",
                    "example": "USD"
                }
            }
        },
...
        "internal_handler_webapp.onRampIntentResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2026-07-02T12:00:00Z"
                },
                "destinationCAddress": {
                    "type": "string",
                    "example": "CABC...XYZ"
                },
                "fiatAmount": {
                    "type": "string",
                    "example": "25"
                },
                "fiatCode": {
                    "type": "string",
                    "example": "USD"
                },
                "id": {
                    "type": "string",
                    "example": "9c1e1c1a-2b3d-4e5f-8a9b-0c1d2e3f4a5b"
                },
                "memoId": {
                    "type": "string",
                    "example": "1234567890"
                },
                "moonpayTransactionId": {
                    "type": "string",
                    "example": "tx_abc123"
                },
                "moonpayTransactionStatus": {
                    "type": "string",
                    "example": "completed"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "created",
                        "pending",
                        "completed",
                        "failed"
                    ],
                    "example": "pending"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2026-07-02T12:05:00Z"
                }
            }
        },
        "internal_handler_webapp.onRampPoolResponse": {
            "type": "object",
            "properties": {
                "network": {
                    "type": "string",
                    "example": "testnet"
                },
                "poolAddress": {
                    "type": "string",
                    "example": "GPOOL...XYZ"
                },
                "recentTransactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_handler_webapp.onRampPoolTransactionResponse"
                    }
                },
                "xlmBalance": {
                    "type": "string",
                    "example": "1234.5000000"
                }
            }
        },
        "internal_handler_webapp.onRampPoolTransactionResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2026-07-02T12:00:00Z"
                },
                "memo": {
                    "type": "string",
                    "example": "1234567890"
                },
                "memoType": {
                    "type": "string",
                    "example": "text"
                },
                "successful": {
                    "type": "boolean",
                    "example": true
                },
                "transactionId": {
                    "type": "string",
                    "example": "a1b2c3..."
                }
            }
        },
        "internal_handler_webapp.onRampSessionResponse": {
            "type": "object",
            "properties": {
                "destinationCAddress": {
                    "type": "string",
                    "example": "CABC...XYZ"
                },
                "fiatAmount": {
                    "type": "string",
                    "example": "25"
                },
                "fiatCode": {
                    "type": "string",
                    "example": "USD"
                },
                "integrationMode": {
                    "type": "string",
                    "enum": [
                        "widget",
                        "platform"
                    ],
                    "example": "widget"
                },
                "intentId": {
                    "type": "string",
                    "example": "9c1e1c1a-2b3d-4e5f-8a9b-0c1d2e3f4a5b"
                },
                "memoId": {
                    "type": "string",
                    "example": "1234567890"
                },
                "platformFallback": {
                    "type": "boolean",
                    "example": true
                },
                "poolAddress": {
                    "type": "string",
                    "example": "GPOOL...XYZ"
                },
                "sessionToken": {
                    "type": "string",
                    "example": "eyJhbGciOi..."
                },
                "widgetUrl": {
                    "type": "string",
                    "example": "https://buy.moonpay.com?apiKey=...\u0026signature=..."
                }
            }
        },
...
        "internal_handler_webapp.updateOnRampIntentRequest": {
            "type": "object",
            "properties": {
                "moonpayTransactionId": {
                    "type": "string",
                    "example": "tx_abc123"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "created",
                        "pending",
                        "completed",
                        "failed"
                    ],
                    "example": "pending"
                }
            }
        },
        "internal_handler_webapp.webappErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "internal_error"
                },
                "error": {
                    "type": "string",
                    "example": "internal error"
                },
                "message": {
                    "type": "string",
                    "example": "internal error"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter: Bearer \u003caccess_token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "RecoveryAuth": {
            "description": "Enter: Bearer \u003crecovery_token\u003e  (obtained from POST /v1/recovery/verify)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}